requirements.txt
# requirements.txt
streamlit==1.37.0
PyMuPDF==1.24.2
transformers==4.43.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
torch==2.3.0
numpy==1.26.4
accelerate==0.32.1
